Added better management of API caller's permissions

+++ app/src/main/java/com/churchclerk/demoapi/DemoApi.java
@@ -4,6 +4,7 @@
 package com.churchclerk.demoapi;
 
 import com.churchclerk.baseapi.BaseApi;
+import com.churchclerk.baseapi.model.ApiCaller;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,14 +12,17 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
 
+import javax.ws.rs.BadRequestException;
+import javax.ws.rs.ForbiddenException;
 import javax.ws.rs.Path;
 import java.util.Date;
+import java.util.UUID;
 
 /**
  *
  */
 @Component
-@Path("/test")
+@Path("/demo")
 public class DemoApi extends BaseApi<Demo> {
 
     private static Logger logger = LoggerFactory.getLogger(DemoApi.class);
@@ -32,9 +36,15 @@ public class DemoApi extends BaseApi<Demo> {
      */
     public DemoApi() {
         super(logger, Demo.class);
-        setReadRoles(Role.ADMIN, Role.CLERK, Role.OFFICIAL, Role.MEMBER, Role.NONMEMBER);
-        setUpdateRoles(Role.ADMIN);
-        setDeleteRoles(Role.ADMIN);
+        setReadRoles(ApiCaller.Role.ADMIN, ApiCaller.Role.CLERK, ApiCaller.Role.OFFICIAL, ApiCaller.Role.MEMBER);
+        setUpdateRoles(ApiCaller.Role.ADMIN);
+        setDeleteRoles(ApiCaller.Role.ADMIN);
+    }
+
+    @Override
+    protected Page<? extends Demo> doGet(Pageable pageable) {
+
+        return service.getResources(pageable, createCriteria());
     }
 
     /**
@@ -46,45 +56,79 @@ public class DemoApi extends BaseApi<Demo> {
 
         addBaseCriteria(criteria);
 
+        if (readAllowed("testmember") == false) {
+            criteria.setId(UUID.randomUUID().toString());
+        }
+
         return criteria;
     }
 
     @Override
-    protected Page<? extends Demo> doGet(Pageable pageable) {
+    protected Demo doGet() {
 
-        return service.getResources(pageable, createCriteria());
-    }
+        if ((id == null) || (id.trim().isEmpty())) {
+            throw new BadRequestException("Resource id cannot be empty");
+        }
 
-    @Override
-    protected Demo doGet(String id) {
+        if (readAllowed(id.trim()) == false) {
+            throw new ForbiddenException();
+        }
 
-            Demo resource = service.getResource(id);
+        Demo resource = service.getResource(id.trim());
 
-            return resource;
+        return resource;
     }
 
     @Override
     protected Demo doCreate(Demo resource) {
 
-        resource.setCreatedBy(requesterId);
+        if (resource.getId() != null) {
+            throw new BadRequestException("Resource id should not be present");
+        }
+
+        if (createAllowed(null) == false) {
+            throw new ForbiddenException();
+        }
+
+        resource.setId(UUID.randomUUID().toString());
+
+        resource.setCreatedBy(apiCaller.getUserid());
         resource.setCreatedDate(new Date());
-        resource.setUpdatedBy(requesterId);
+        resource.setUpdatedBy(apiCaller.getUserid());
         resource.setUpdatedDate(new Date());
         return resource;
     }
 
     @Override
     protected Demo doUpdate(Demo resource) {
-        Demo found = service.getResource(id);
+        if ((id == null) || id.isEmpty() || resource.getId() == null || resource.getId().isEmpty()) {
+            throw new BadRequestException("Resource id cannot be empty");
+        }
 
-        resource.setUpdatedBy(requesterId);
+        if (resource.getId().equals(id) == false) {
+            throw new BadRequestException("Resource id does not match");
+        }
+
+        if (updateAllowed(id) == false) {
+            throw new ForbiddenException();
+        }
+
+        resource.setUpdatedBy(apiCaller.getUserid());
         resource.setUpdatedDate(new Date());
 
         return service.updateResource(resource);
     }
 
     @Override
-    protected Demo doDelete(String id) {
+    protected Demo doDelete() {
+        if ((id == null) || id.isEmpty()) {
+            throw new BadRequestException("Resource id cannot be empty");
+        }
+
+        if (deleteAllowed(id) == false) {
+            throw new ForbiddenException();
+        }
+
         return service.deleteResource(id);
     }
 }
